[LVLUPDB-230] Multicast listener using RxJava (#1)
This commit adds an implementation of `SpyqlListener` that extends RxJava's `Observable`. All callbacks of `SpyqlListener` are encoded as `Events` and pushed for processing downstream. This `Observable` API can be treated as a low-level API.
In addition to that there is a high-level API provided by `ObservableListerner.attach[Async]Listener()` methods that allows to multicast events to multiple `SpyqlListener`s.